fix(ci): rename the composable and move the file
